+++ models/owner.js
@@ -33,7 +33,7 @@ module.exports = (sequelize, DataTypes) => {
         allowNull: false,
       },
       phone: {
-        type: DataTypes.INTEGER(20),
+        type: DataTypes.BIGINT(),
         allowNull: false,
       },
       address: {
